requirements.txt
# Core Framework Dependencies
pymongo==3.12.3
motor==2.5.1
python-dotenv==1.0.0
faker==20.1.0
pandas==2.1.4
numpy==1.24.3
tqdm==4.66.1
aiofiles==23.2.1
pydantic==2.5.2
dnspython>=1.16.0
uvloop==0.19.0
zstandard==0.25.0

# Framework Additional Dependencies
psutil>=5.9.0
pyyaml>=6.0
asyncio-mqtt>=0.11.0
